requirements_bot.txt
# Dependências do Bot Telegram com Supabase

# Pyrogram - Framework do bot Telegram
pyrogram==1.4.16
TgCrypto==1.2.5

# Supabase - Cliente Python
supabase==2.4.1
python-dotenv==1.0.0

# Utilitários
requests==2.31.0
python-dateutil==2.8.2
